# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BhuSetu",
    "description": "BhuSetu: Intelligent Land Record Digitization and Validation System for SIH Problem Statement 26018 (Dept of Land Resources, Govt of India)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/geo/crosscheck": {
      "get": {
        "tags": [
          "Geospatial"
        ],
        "summary": "Cross Check Survey",
        "description": "Cross-validates document survey area against cadastral GIS parcel boundary.",
        "operationId": "cross_check_survey_api_geo_crosscheck_get",
        "parameters": [
          {
            "name": "survey_no",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Survey or Khasra Number",
              "title": "Survey No"
            },
            "description": "Survey or Khasra Number"
          },
          {
            "name": "village",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Village Name",
              "default": "",
              "title": "Village"
            },
            "description": "Village Name"
          },
          {
            "name": "textual_area_sqm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Recorded textual area in sq metres",
              "title": "Textual Area Sqm"
            },
            "description": "Recorded textual area in sq metres"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/geo/crosscheck": {
      "get": {
        "tags": [
          "Geospatial"
        ],
        "summary": "Cross Check Survey",
        "description": "Cross-validates document survey area against cadastral GIS parcel boundary.",
        "operationId": "cross_check_survey_api_v1_geo_crosscheck_get",
        "parameters": [
          {
            "name": "survey_no",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Survey or Khasra Number",
              "title": "Survey No"
            },
            "description": "Survey or Khasra Number"
          },
          {
            "name": "village",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Village Name",
              "default": "",
              "title": "Village"
            },
            "description": "Village Name"
          },
          {
            "name": "textual_area_sqm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Recorded textual area in sq metres",
              "title": "Textual Area Sqm"
            },
            "description": "Recorded textual area in sq metres"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}